requirements.txt
click==8.1.7
colorama==0.4.6
Flask==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2==2.9.6
Werkzeug==2.0.1
python-dotenv==1.0.0
bcrypt==4.0.1
blinker==1.6.2
